Relocating ReflectUtility class into repicea.serial package. Removing deprecated classes.
